Add tools for getting and viewing mnist data
The mnist data is available on a well known mirror.  This allows
developers to quickly fetch that data without checking it in.

A simple tool is also provided as an exploration of the data set
that can convert the idx data into png files for inspecting a given
training or test example